Removed useless fields from grow edit, Reactified the data graphs
